refactor : move commands from service to domain
