+++ src/main/java/nextstep/subway/controller/request/StationCreateRequest.java
@@ -1,6 +1,6 @@
 package nextstep.subway.controller.request;
 
-import nextstep.subway.service.command.StationCreateCommand;
+import nextstep.subway.domain.command.StationCreateCommand;
 
 public class StationCreateRequest implements StationCreateCommand {
     private String name;
