+++ src/main/java/nextstep/subway/domain/command/LineModifyCommand.java
@@ -1,4 +1,4 @@
-package nextstep.subway.service.command;
+package nextstep.subway.domain.command;
 
 public interface LineModifyCommand {
     String getName();
